Refac: charges_reader_tests.
